add working form to RecapVideoMaker.vue, rename videoMaker
