+++ components/videoMaker/RecapVideoMaker.vue
@@ -10,7 +10,7 @@
     <div class="field">
       <div class="control">
         <label class="checkbox">
-          <input type="checkbox">
+          <input type="checkbox" v-model="hasAcceptedCondition">
             J'accepte que le film soit mis en ligne et <a href="#">accepte les conditions d'utilisations</a>.
         </label>
       </div>
@@ -18,18 +18,36 @@
 
     <div class="field is-grouped">
       <div class="control">
-        <button class="button is-link">Valider</button>
+        <button class="button is-link" :disabled="!this.isValid" @click="sendNewVideoToApi(inputContent)">Valider</button>
       </div>
     </div>
   </div>
 </template>
 
 <script>
+import {mapActions} from 'vuex';
+
 export default {
   name: "RecapVideoMaker",
   data() {
     return {
-      inputContent: ''
+      inputContent: '',
+      hasAcceptedCondition: false
+    }
+  },
+  methods: {
+    ...mapActions({
+      sendNewVideoToApi: 'SEND_NEW_MOVIE'
+    })
+  },
+  computed: {
+    /**
+     * Si le titre est plus long que 3 caractères et
+     * que l'utilisateur a accepter les conditions d'utilisations
+     * @returns {boolean}
+     */
+    isValid: function() {
+      return this.inputContent.length > 3 && this.hasAcceptedCondition;
     }
   }
 }
