+++ package.json
@@ -28,6 +28,7 @@
     "jest": "^26.4.2",
     "node-sass": "^4.14.1",
     "sass-loader": "^9.0.3",
+    "vue": "^2.6.12",
     "vue-jest": "^3.0.4",
     "vuex": "^3.5.1"
   }
