+++ store/index.js
@@ -66,10 +66,10 @@ export const actions = {
    * @param commit
    * @constructor
    */
-  GET_CHAPTER_VIDEOS_LIST: function({ state, commit }) {
+  GET_CHAPTER_VIDEOS_LIST: async function({ state, commit }) {
     commit('SET_LOADING', true);
     //TODO Changer l'URL de l'API pour y inclure le chapitre actuel en le récupérant depuis le state
-    this.$axios.$get(`https://picsum.photos/v2/list?limit=10`)
+    await this.$axios.$get(`https://picsum.photos/v2/list?limit=10`)
       .then(response => {
         commit('SET_CHAPTER_VIDEOS_LIST', { list: response });
         commit('SET_LOADING', false);
@@ -78,19 +78,20 @@ export const actions = {
       })
     ;
   },
-
   /**
    * Envoie une nouvelle vidéo vers l'API
    * @param commit
+   * @param payload
    * @constructor
    */
-  SEND_NEW_MOVIE: function({commit}, payload) {
+  SEND_NEW_MOVIE: async function({commit}, payload) {
     //TODO SEND NEW FILM DATA TO API
-    this.$axios.$post('')
-      .then(r => {
 
-      }).catch(r => {
-
-    });
+    // this.$axios.$post('')
+    //   .then(r => {
+    //
+    //   }).catch(r => {
+    //
+    // });
   }
 }
